requirements.txt
tensorflow-gpu==2.8.0
librosa==0.9.1
matplotlib==3.5.1
numpy==1.21.5
tqdm~=4.62.3
keras~=2.8.0
keras-tcn
audioread